+++ server/src/test/java/org/opensearch/extensions/ExtensionsManagerTests.java
@@ -187,7 +187,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
 
         Set<Setting<?>> additionalSettings = extAwarePlugin.getExtensionSettings().stream().collect(Collectors.toSet());
         ExtensionScopedSettings extensionScopedSettings = new ExtensionScopedSettings(additionalSettings);
-        ExtensionsManager extensionsManager = new ExtensionsManager(additionalSettings);
+        ExtensionsManager extensionsManager = new ExtensionsManager(additionalSettings, identityService);
         ExtensionDependency dependentExtension = new ExtensionDependency("uniqueid0", Version.fromString("2.0.0"));
 
         Extension firstExtension = new Extension(
@@ -283,7 +283,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
             null,
             null
         );
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         extensionsManager.loadExtension(firstExtension);
         IOException exception = expectThrows(IOException.class, () -> extensionsManager.loadExtension(secondExtension));
         assertEquals(
@@ -324,7 +324,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     public void testMissingRequiredFieldsWhileLoadingExtension() throws Exception {
 
         Extension firstExtension = new Extension("firstExtension", "uniqueid1", "127.0.0.0", "9300", "0.0.7", "3.0.0", "", null, null);
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
 
         IOException exception = expectThrows(IOException.class, () -> extensionsManager.loadExtension(firstExtension));
         assertEquals("Required field [minimum opensearch version] is missing in the request", exception.getMessage());
@@ -381,7 +381,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testInitialize() throws Exception {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
 
         initialize(extensionsManager);
 
@@ -493,7 +493,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
 
     public void testHandleRegisterRestActionsRequest() throws Exception {
 
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         initialize(extensionsManager);
 
         String uniqueIdStr = "uniqueid1";
@@ -564,7 +564,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testHandleRegisterSettingsRequest() throws Exception {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         initialize(extensionsManager);
 
         String uniqueIdStr = "uniqueid1";
@@ -580,7 +580,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testHandleRegisterRestActionsRequestWithInvalidMethod() throws Exception {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         initialize(extensionsManager);
 
         String uniqueIdStr = "uniqueid1";
@@ -598,7 +598,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testHandleRegisterRestActionsRequestWithInvalidDeprecatedMethod() throws Exception {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         initialize(extensionsManager);
 
         String uniqueIdStr = "uniqueid1";
@@ -616,7 +616,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testHandleRegisterRestActionsRequestWithInvalidUri() throws Exception {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         initialize(extensionsManager);
         String uniqueIdStr = "uniqueid1";
         extensionsManager.loadExtension(
@@ -633,7 +633,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testHandleRegisterRestActionsRequestWithInvalidDeprecatedUri() throws Exception {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         initialize(extensionsManager);
         String uniqueIdStr = "uniqueid1";
         extensionsManager.loadExtension(
@@ -650,7 +650,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testHandleExtensionRequest() throws Exception {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         initialize(extensionsManager);
 
         ExtensionRequest clusterStateRequest = new ExtensionRequest(ExtensionRequestProto.RequestType.REQUEST_EXTENSION_CLUSTER_STATE);
@@ -804,7 +804,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testAddSettingsUpdateConsumerRequest() throws Exception {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         initialize(extensionsManager);
 
         List<Setting<?>> componentSettings = List.of(
@@ -848,7 +848,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testHandleAddSettingsUpdateConsumerRequest() throws Exception {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         initialize(extensionsManager);
 
         List<Setting<?>> componentSettings = List.of(
@@ -868,7 +868,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testUpdateSettingsRequest() throws Exception {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         initialize(extensionsManager);
 
         Setting<?> componentSetting = Setting.boolSetting("falseSetting", false, Property.Dynamic);
@@ -897,7 +897,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
 
     public void testRegisterHandler() throws Exception {
 
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
 
         TransportService mockTransportService = spy(
             new TransportService(
@@ -925,7 +925,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
     }
 
     public void testIncompatibleExtensionRegistration() throws IOException {
-        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of());
+        ExtensionsManager extensionsManager = new ExtensionsManager(Set.of(), identityService);
         Extension firstExtension = new Extension(
             "firstExtension",
             "uniqueid1",
@@ -966,7 +966,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
             extensionScopedSettings
         );
 
-        ExtensionsManager extensionsManager = new ExtensionsManager(additionalSettings);
+        ExtensionsManager extensionsManager = new ExtensionsManager(additionalSettings, identityService);
         extensionsManager.loadExtension(firstExtension);
 
         DiscoveryExtensionNode extension = new DiscoveryExtensionNode(
@@ -1004,7 +1004,7 @@ public class ExtensionsManagerTests extends OpenSearchTestCase {
             extensionScopedSettings
         );
 
-        ExtensionsManager extensionsManager = new ExtensionsManager(additionalSettings);
+        ExtensionsManager extensionsManager = new ExtensionsManager(additionalSettings, identityService);
         extensionsManager.loadExtension(firstExtension);
 
         DiscoveryExtensionNode extension = new DiscoveryExtensionNode(